Adds warning message to assigning auto track
The user must now confirm that they will be overwriting previous points
during the frames listed on the auto track. If cancelled, a notification
will pop up notifying them that the merge did not occur.

***Note that there is an issue when assigning at the moment. It will
throw an exception at the moment if no points have been manually chosen.
As it is attempting to output the total points listed...